requirements.txt
Pillow-7.1.2
